AnalystTool/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
openai
apscheduler